# Frame sync changes for higher performance & fps