--- a/neo/framework/File_SaveGame.cpp
+++ b/neo/framework/File_SaveGame.cpp
@@ -71,7 +71,9 @@ class idSGFcompressThread : public idSysThread
 public:
 	virtual int			Run()
 	{
+#if USE_OPTICK
 		OPTICK_THREAD( "idSGFcompressThread" );
+#endif
 
 		sgf->CompressBlock();
 		return 0;
@@ -83,7 +85,9 @@ class idSGFdecompressThread : public idSysThread
 public:
 	virtual int			Run()
 	{
+#if USE_OPTICK
 		OPTICK_THREAD( "idSGFdecompressThread" );
+#endif
 
 		sgf->DecompressBlock();
 		return 0;
@@ -95,7 +99,9 @@ class idSGFwriteThread : public idSysThread
 public:
 	virtual int			Run()
 	{
+#if USE_OPTICK
 		OPTICK_THREAD( "idSGFwriteThread" );
+#endif
 
 		sgf->WriteBlock();
 		return 0;
@@ -107,7 +113,9 @@ class idSGFreadThread : public idSysThread
 public:
 	virtual int			Run()
 	{
+#if USE_OPTICK
 		OPTICK_THREAD( "idSGFreadThread" );
+#endif
 
 		sgf->ReadBlock();
 		return 0;