--- a/neo/idlib/precompiled.h
+++ b/neo/idlib/precompiled.h
@@ -107,7 +107,7 @@ const int MAX_EXPRESSION_REGISTERS = 4096;
 #endif
 
 // RB: make Optick profiling available everywhere
-#if defined( USE_OPTICK )
+#if USE_OPTICK
 	#include "../libs/optick/optick.h"
 #endif