--- a/neo/sys/DeviceManager.h
+++ b/neo/sys/DeviceManager.h
@@ -61,7 +61,6 @@ struct DeviceCreationParameters
 	nvrhi::Format swapChainFormat = nvrhi::Format::RGBA8_UNORM; // RB: don't do the sRGB gamma ramp with the swapchain
 	uint32_t swapChainSampleCount = 1;
 	uint32_t swapChainSampleQuality = 0;
-	uint32_t maxFramesInFlight = 2;
 	bool enableDebugRuntime = false;
 	bool enableNvrhiValidationLayer = false;
 	bool vsyncEnabled = false;