--- a/neo/sys/DeviceManager_DX12.cpp
+++ b/neo/sys/DeviceManager_DX12.cpp
@@ -61,6 +61,7 @@ class DeviceManager_DX12 : public DeviceManager
 	std::vector<nvrhi::TextureHandle>           m_RhiSwapChainBuffers;
 	RefCountPtr<ID3D12Fence>                    m_FrameFence;
 	std::vector<HANDLE>                         m_FrameFenceEvents;
+	nvrhi::EventQueryHandle						m_FrameWaitQuery;
 
 	UINT64                                      m_FrameCount = 1;
 
@@ -447,6 +448,9 @@ bool DeviceManager_DX12::CreateDeviceAndSwapChain()
 		m_FrameFenceEvents.push_back( CreateEvent( nullptr, false, true, NULL ) );
 	}
 
+	m_FrameWaitQuery = nvrhiDevice->createEventQuery();
+	nvrhiDevice->setEventQuery( m_FrameWaitQuery, nvrhi::CommandQueue::Graphics );
+
 	return true;
 }
 
@@ -458,6 +462,8 @@ void DeviceManager_DX12::DestroyDeviceAndSwapChain()
 	ReleaseRenderTargets();
 
 	nvrhiDevice = nullptr;
+
+	m_FrameWaitQuery = nullptr;
 
 	for( auto fenceEvent : m_FrameFenceEvents )
 	{
@@ -647,6 +653,11 @@ void DeviceManager_DX12::Present()
 
 	// SRS - Don't change deviceParms.vsyncEnabled here, simply test for vsync mode 2 to set DXGI SyncInterval
 	m_SwapChain->Present( deviceParms.vsyncEnabled && r_swapInterval.GetInteger() == 2 ? 1 : 0, presentFlags );
+
+	// SRS - Sync on previous frame's command queue completion vs. waitForIdle() on whole device
+	nvrhiDevice->waitEventQuery( m_FrameWaitQuery );
+	nvrhiDevice->resetEventQuery( m_FrameWaitQuery );
+	nvrhiDevice->setEventQuery( m_FrameWaitQuery, nvrhi::CommandQueue::Graphics );
 
 	m_FrameFence->SetEventOnCompletion( m_FrameCount, m_FrameFenceEvents[bufferIndex] );
 	m_GraphicsQueue->Signal( m_FrameFence, m_FrameCount );