--- a/neo/sys/common/savegame.cpp
+++ b/neo/sys/common/savegame.cpp
@@ -746,7 +746,9 @@ idSaveGameThread::Run
 */
 int idSaveGameThread::Run()
 {
+#if USE_OPTICK
 	OPTICK_THREAD( "idSaveGameThread" );
+#endif
 
 	int ret = ERROR_SUCCESS;