Over the past few weeks I have been doing some profiling using Xcode Instruments to see if I could improve macOS performance / fps.  This process allowed me to observe that RBDoom3BFG is largely serializing CPU-side command queue processing and GPU processing.  This is especially bad on macOS given that Vulkan commands have to be translated via MoltenVK to Metal commands on the fly, and this overhead makes things worse compared to Windows and Linux where Vulkan is native.  The main reason for the serialization is due to the device-level `waitForIdle()` call inside `GL_BlockingSwapBuffers()`, where the CPU waits for GPU operations to complete before getting started on the next frame's game thread and command queue processing.  This limits the parallel overlap between the CPU and GPU and really slows things down.  For vanilla versions of Doom3 this may not be so bad since rendering was fairly simple.  However, with RBDoom3BFG rendering is becoming more complex and time consuming, and parallelism is needed.

I noticed that NVRHI offers other modes of sync, most notably command queue event queries.  This pull request experiments with two things: a) using triple frame buffering vs. double buffering for NVRHI (DX12 & Vulkan), and b) replacing the device `waitForIdle()` device sync with a command queue sync operation based on completion of the previous frame's graphics command queue.  The results with my AMD 6600XT card are fairly significant as seen below:

Win32 DX12: before: 50 fps (as of Feb 28, before additional commits), after: 136 fps
Win32 Vulkan: before: 74 fps (as of today using up-to-date master), after: 133 fps
Linux Vulkan: before: 27 fps (as of Feb 28, before additional commits), after: 122 fps
macOS Vulkan: before: 38 fps (as of today using up-to-date master), after: 58 fps

This work has been tested using my AMD 6600XT card on all OSs, as well as an internal Intel UHD 630 graphics processor (on my Intel 8700K) to test these sync changes on a slow GPU.  So far so good.  However, I would like to see how these operate on a faster GPU (i.e. a more powerful Nvidia or AMD chip).

In addition this pull request also includes:

1) Save screenshots to **fs_savepath** (as per original Doom3 default) vs. **fs_basepath**
2) Fix SDL display mode size/refresh rate changes when already in fullscreen mode
3) Fix compilation errors when Optick profiling is OFF.  (note that setting OPTICK=ON for macOS/linux exposes some issues with including optick.h in precompiled.h, but this may not be the main use case so I did not attempt to fix.  On macOS the default profiler is Instruments' Game Performance tool)

Win32 DX12 before and after screenshots:

![win32-dx12-before-pic1](https://user-images.githubusercontent.com/82544213/222941824-b607031e-0317-4331-9670-9aefbb3e2775.png)
![win32-dx12-after](https://user-images.githubusercontent.com/82544213/222941841-f2ddd4d4-e6bc-4d7c-8579-6e9d23a83ee6.png)

Win32 Vulkan before and after screenshots:

![win32-vulkan-before](https://user-images.githubusercontent.com/82544213/222941858-ad5c4995-f316-449d-9e57-84bc2453980a.png)
![win32-vulkan-after](https://user-images.githubusercontent.com/82544213/222941865-8741c742-8c16-4c84-9265-5b6dd6cfd5f6.png)