@RobertBeckebans 2023-03-05
This is very interesting takes some time for reviewing. The BFG edition has already an improved syncing behaviour compared to vanilla Doom 3 that shifts the GPU work into the next CPU frame.
In vanilla D3 it was just BeginFrame() -> build all GL commands, wait for completion as in EndFrame() and then swap the buffers immediatly. I also integrated Optick so we can see better what's happening concerning this.

Could you remove the screenshot path change? Vanilla D3 has savepath but it wasn't used on Windows and in general it is just annoying to only go to the home savepath just for sharing screenshots. Especially if you work on a mod you just want them drag out of your mod working directory. So that change was intentional in this source port.