@RobertBeckebans 2023-03-05
Also the Optick related changes seem to be wrong.
The optick.h should be always included in precompiled.h and then
`
if(OPTICK)
	target_compile_definitions(RBDoom3BFG PUBLIC USE_OPTICK=1)
else()
	target_compile_definitions(RBDoom3BFG PUBLIC USE_OPTICK=0)
endif()
`
sets USE_OPTICK. So all Optick macros don't do anything if cmake -DOPTICK=OFF and don't need an extra #ifdef USE_OPTICK wrap.