@RobertBeckebans 2023-03-05
Sorry I haven't tested Optick on Linux before merging it into master. The last commit fixes the compilation but requires a rebase on your PR.