@RobertBeckebans 2023-03-05
Holy shit this is a speed increase from 143 fps to 240 on my AMD laptop.
 
![rbdoom-3-bfg-20230305-123920-001](https://user-images.githubusercontent.com/1278614/222958348-80651fc3-52d4-472a-bac7-77ba2811bc59.png)