@SRSaunders 2023-03-05
Thanks for looking at this PR.

1. I will revert the screenshot change for Windows and Linux.  On macOS putting the screenshots into basepath is pretty bad since it may be located inside of an app bundle.
2. I will also rebase the PR onto your Optick fix and test again for compilation on macOS / linux.
3. I will add one small revert for compileshaders.cmake to bring macOS up to the SPRIV target = vulkan1.2
4. I would also like to revert the parallel shader compilation test for macOS, but this depends on a pending PR in your nvrhi clone.  I did notice that you have a pending PR on updating nvrhi, but are also facing parallel compilation problems on Windows.  Does my pending nvrhi tools fix help in any way?

Mainly I am interested in your review of the perf improvement and if it's a safe change that meets sync requirements.