@SRSaunders 2023-03-05
Changes completed now.  Thanks for the feedback on USE_OPTICK.  I was not aware of the optick.config.h file and the setting in there was causing problems on linux/macOS.