@RobertBeckebans 2023-03-06
One of the worst case scenes in Doom 3 with up to 50 point lights and many of them cast shadows:
All shots in 1080p

RBDOOM-3-BFG 1.4.0 release binaries with OpenGL 91 fps

![rbdoom-3-bfg-20230306-121203-001](https://user-images.githubusercontent.com/1278614/223095836-3473f355-ce2f-40d0-a5b4-e906df9e9256.png)

The master branch has few less draw calls because it fades out distant shadow casting lights and the shadowmap resolution is only half of 1.4.0 so all shadows fit into the 8192^2 atlas.

master branch with DX12 107 fps
![rbdoom-3-bfg-20230306-121252-001](https://user-images.githubusercontent.com/1278614/223095926-240c3b5c-fee1-46b7-9782-606142f63c7f.png)

sync-changes branch with DX12 176 fps
![rbdoom-3-bfg-20230306-121330-001](https://user-images.githubusercontent.com/1278614/223095981-361099ca-6b5d-4a15-9742-1d5c7cfb6259.png)

master branch with Vulkan 70 fps
![rbdoom-3-bfg-20230306-122203-001](https://user-images.githubusercontent.com/1278614/223097008-b09c7b7e-cceb-48eb-b69a-033622c1bd38.png)

sync-changes branch with Vulkan 111 fps
![rbdoom-3-bfg-20230306-122244-001](https://user-images.githubusercontent.com/1278614/223097080-66311d4a-f66e-4eb0-867f-12052da59a52.png)

I also brought this up to discussion in Discord and these are the results by Samson (KozGit DOOM-3-BFG VR author):

master branch with DX12 429 fps
![rbdoom-3-bfg-20230305-155412-001](https://user-images.githubusercontent.com/1278614/223097616-e7fcf3fc-7d1f-4ca7-b172-50c07efd9e80.png)

sync-changes branch with DX12 ... 666 (teleporter to hell speed) fps
![RBDoom3BFG_sync_2023_03_05_23_30_38_770](https://user-images.githubusercontent.com/1278614/223098013-ebb89fb5-91a7-4f5f-8cf2-4268ccfdf43d.png)