@RobertBeckebans 2023-03-06
I don't understand why this requires NUM_FRAME_DATA = 3. If I understand this correctly then there should be only a single synchronization point between 2 frames in GL_BlockingSwapBuffers.

Is it at the moment that `
void DeviceManager_DX12::Present()
{
	// SRS - Sync on previous frame's command queue completion vs. waitForIdle() on whole device
	m_NvrhiDevice->waitEventQuery( m_FrameWaitQuery );
	m_NvrhiDevice->resetEventQuery( m_FrameWaitQuery );
	m_NvrhiDevice->setEventQuery( m_FrameWaitQuery, nvrhi::CommandQueue::Graphics );
`
doesn't wait for the completion of the previous frame's NVHRI commandlist but the one second to last?
I get a lot of flickering when I try NUM_FRAME_DATA  = 2 on Windows.

It seems that GL_BlockingSwapBuffers became much faster during profiles:
![2023-03-06 20_20_26-What the hell is going on_ - Optick Profiler](https://user-images.githubusercontent.com/1278614/223210411-f14b4a09-1d37-4f18-bad0-45d2d7f9869c.jpg)