@RobertBeckebans 2023-03-06
This is the frame overlap with r_swapInterval 0 and com_fixedTic 1 to go over 120 Hz:

![2023-03-06 23_07_11-What the hell is going on_ - Optick Profiler](https://user-images.githubusercontent.com/1278614/223249555-43525baf-49dd-4b30-bebd-eb3874591d8c.png)


The required changes are in branch 750-better-vsync.