@SRSaunders 2023-03-07
Thanks @RobertBeckebans for stress testing this possible change under various scenarios, and especially for reaching out on Discord for testing on a fast RTX 4090 rig.  I believe NUM_FRAME_DATA = 3 helps by increasing the availability of free swap chain images as the frame rate increases (but with constant rendering work per frame).

Regarding the flickering with NUM_FRAME_DATA = 2, I am concerned that I may not have implemented the correct sync operation to guarantee proper results under the more stressful condition of fewer swap chain images, i.e. double vs. triple buffering without `waitForIdle()` slowing things down.  At least on Vulkan, I have verified that I implemented a wait on the barrier command list which terminates a frame.  And I do believe the wait is on the previous frame's command list and not 2 frames back.

If @stephenap07 has an alternative sync idea with fences that could survive NUM_FRAME_DATA = 2 without image corruption I would be happy to see that improvement.  As an aside, I played around with this concept on macOS with MoltenVK using the fence in `vkAcquireNextImageKHR()`, but could not get any improvement over what I did with nvrhi queue events.  Perhaps I was doing something wrong so I am curious to see if Stephen P has a better solution here.